pyproject.toml
[tool.poetry]
name = "air-django"
version = "0.1.0"
description = ""
authors = ["knw <kimnamwook08@gmail.com>"]
readme = "README.md"
packages = [{include = "air_django"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
djangorestframework = "^3.14.0"
strawberry-graphql-django = "^0.9.4"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
